Few things about LNS:
- update random neighborhood: nbFixedVars is now a double,
- add fast restart direct in LNS, instead of external declaration, thanks to a counter,
- add a parameter to SearchLoop#interrupt(String message)

+++ choco-solver/src/main/java/solver/search/loop/lns/neighbors/RandomNeighborhood.java
@@ -47,7 +47,7 @@ public class RandomNeighborhood implements INeighbor {
     private final IntVar[] vars;
     private final int[] bestSolution;
     private Random rd;
-    private int nbFixedVariables = 0;
+    private double nbFixedVariables = 0d;
 
     BitSet fragment;  // index of variable to set unfrozen
